TLAR
#410

Added a requirementType and changed all values in aircraft/global to
this type. RequirementType indicates a required value and an actual
value.

There are still several issues open for discussion:

1. Currently, all values are set to be double. At least boolean is
needed, e.g. The aircraft is required to fulfill CS-25 and there must be
a true/false identification to check this requirement.

2. The parameters are currently grouped according to the setup in an
internal project. This may be improved, e.g. move designRange to the
performanceTargets gruop

3. Documentation needs to be added before issue can be closed

4. We are missing an indication on whether a requirement is satisfied if
the value is bigger/greater/equal. Maybe add a string enumeration to the
requirementsType.